requirements.txt
odf
pymupdf
pillow
numpy
base64
